+++ file_lists/source_files_no_main
@@ -2,7 +2,6 @@ src/stim.cc
 src/stim/arg_parse.cc
 src/stim/circuit/circuit.cc
 src/stim/circuit/circuit_instruction.cc
-src/stim/circuit/export_qasm.cc
 src/stim/circuit/gate_decomposition.cc
 src/stim/circuit/gate_target.cc
 src/stim/cmd/command_analyze_errors.cc
@@ -92,4 +91,5 @@ src/stim/util_bot/error_decomp.cc
 src/stim/util_top/circuit_inverse_unitary.cc
 src/stim/util_top/circuit_to_detecting_regions.cc
 src/stim/util_top/circuit_vs_amplitudes.cc
+src/stim/util_top/export_qasm.cc
 src/stim/util_top/simplified_circuit.cc
