+++ file_lists/test_files
@@ -1,7 +1,6 @@
 src/stim.test.cc
 src/stim/arg_parse.test.cc
 src/stim/circuit/circuit.test.cc
-src/stim/circuit/export_qasm.test.cc
 src/stim/circuit/gate_decomposition.test.cc
 src/stim/circuit/gate_target.test.cc
 src/stim/cmd/command_analyze_errors.test.cc
@@ -87,6 +86,7 @@ src/stim/util_top/circuit_inverse_unitary.test.cc
 src/stim/util_top/circuit_to_detecting_regions.test.cc
 src/stim/util_top/circuit_vs_amplitudes.test.cc
 src/stim/util_top/circuit_vs_tableau.test.cc
+src/stim/util_top/export_qasm.test.cc
 src/stim/util_top/simplified_circuit.test.cc
 src/stim/util_top/stabilizers_to_tableau.test.cc
 src/stim/util_top/stabilizers_vs_amplitudes.test.cc
