+++ src/stim.h
@@ -6,7 +6,6 @@
 #include "stim/arg_parse.h"
 #include "stim/circuit/circuit.h"
 #include "stim/circuit/circuit_instruction.h"
-#include "stim/circuit/export_qasm.h"
 #include "stim/circuit/gate_decomposition.h"
 #include "stim/circuit/gate_target.h"
 #include "stim/cmd/command_analyze_errors.h"
@@ -113,6 +112,7 @@
 #include "stim/util_top/circuit_to_detecting_regions.h"
 #include "stim/util_top/circuit_vs_amplitudes.h"
 #include "stim/util_top/circuit_vs_tableau.h"
+#include "stim/util_top/export_qasm.h"
 #include "stim/util_top/simplified_circuit.h"
 #include "stim/util_top/stabilizers_to_tableau.h"
 #include "stim/util_top/stabilizers_vs_amplitudes.h"
