+++ src/stim/circuit/circuit.pybind.cc
@@ -18,7 +18,7 @@
 
 #include "stim/circuit/circuit_instruction.pybind.h"
 #include "stim/circuit/circuit_repeat_block.pybind.h"
-#include "stim/circuit/export_qasm.h"
+#include "stim/util_top/export_qasm.h"
 #include "stim/circuit/gate_target.pybind.h"
 #include "stim/cmd/command_diagram.pybind.h"
 #include "stim/dem/detector_error_model_target.pybind.h"
